+++ src/main/java/com/wasu/model/InventoryHistory.java
@@ -21,9 +21,9 @@ public class InventoryHistory {
 
     private Long inventorystate;
 
-    private Long latitude;
+    private Double latitude;
 
-    private Long longitude;
+    private Double longitude;
 
     private String name;
 
@@ -34,7 +34,7 @@ public class InventoryHistory {
     public InventoryHistory() {
     }
 
-    public InventoryHistory(Long id, String address, Date inventorydate, Long latitude, Long longitude, String photo) {
+    public InventoryHistory(Long id, String address, Date inventorydate, Double latitude, Double longitude, String photo) {
         this.id = id;
         this.address = address;
         this.inventorydate = inventorydate;
@@ -115,19 +115,19 @@ public class InventoryHistory {
         this.inventorystate = inventorystate;
     }
 
-    public Long getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Long latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
 
-    public Long getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Long longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
